Remove draft for sub-page for topics
The intention for sub-pages for the individual pages was to make
it super-easy for developers to find snippets for copy+paste development.

However, it is totally an option to create book-sized documents on a single page, e.g.:
- https://hsc.aim42.org/documentation/hsc_arc42.html
- https://ci.apache.org/projects/wicket/guide/9.x/single.html

Having big documents will make the maintenance of the documentation documents simpler.

So, the architecture and basic technology documentation will be provided in rather big,
comprehensive documents. Potential contributes will need to be able to consume content
which is not bite-sized.